order cancellation implemented (#592)
* order cancellation implemented

* fixed code smells

* fixed bug

* added unregistered service

* Update src/Altinn.Notifications.Persistence/Repository/OrderRepository.cs


* -> canCancel

* removed task from analysis

---------

+++ test/Altinn.Notifications.IntegrationTests/Notifications.Persistence/OrderRepositoryTests.cs
@@ -2,6 +2,7 @@
 using Altinn.Notifications.Core.Models.NotificationTemplate;
 using Altinn.Notifications.Core.Models.Orders;
 using Altinn.Notifications.Core.Persistence;
+using Altinn.Notifications.Core.Shared;
 using Altinn.Notifications.IntegrationTests.Utils;
 using Altinn.Notifications.Persistence.Repository;
 
@@ -25,8 +26,11 @@ namespace Altinn.Notifications.IntegrationTests.Notifications.Persistence
 
         public async Task DisposeAsync()
         {
-            string deleteSql = $@"DELETE from notifications.orders o where o.alternateid in ('{string.Join("','", _orderIdsToDelete)}')";
-            await PostgreUtil.RunSql(deleteSql);
+            if (_orderIdsToDelete.Count != 0)
+            {
+                string deleteSql = $@"DELETE from notifications.orders o where o.alternateid in ('{string.Join("','", _orderIdsToDelete)}')";
+                await PostgreUtil.RunSql(deleteSql);
+            }
         }
 
         [Fact]
@@ -142,6 +146,101 @@ namespace Altinn.Notifications.IntegrationTests.Notifications.Persistence
 
                 Assert.Equal(1, orderCount);
             }
+        }
+
+        [Fact]
+        public async Task CancelOrder_OrderDoesNotExits_ReturnsCancellationError()
+        {
+            // Arrange
+            OrderRepository repo = (OrderRepository)ServiceUtil
+                .GetServices(new List<Type>() { typeof(IOrderRepository) })
+                .First(i => i.GetType() == typeof(OrderRepository));
+
+            // Act
+            Result<NotificationOrderWithStatus, CancellationError> result = await repo.CancelOrder(Guid.NewGuid(), "non-exitent-org");
+
+            // Assert   
+            result.Match(
+                success =>
+                    throw new Exception("No success value should be returned if order is not found in database."),
+                error =>
+                {
+                    Assert.Equal(CancellationError.OrderNotFound, error);
+                    return true;
+                });
+        }
+
+        [Fact]
+        public async Task CancelOrder_SendTimePassed_ReturnsError()
+        {
+            // Arrange
+            OrderRepository repo = (OrderRepository)ServiceUtil
+                .GetServices(new List<Type>() { typeof(IOrderRepository) })
+                .First(i => i.GetType() == typeof(OrderRepository));
+
+            NotificationOrder order = new()
+            {
+                Id = Guid.NewGuid(),
+                Created = DateTime.UtcNow,
+                Creator = new("test"),
+                Templates = new List<INotificationTemplate>()
+                {
+                    new SmsTemplate("Altinn", "This is the body")
+                },
+                RequestedSendTime = DateTime.UtcNow.AddMinutes(-1)
+            };
+
+            _orderIdsToDelete.Add(order.Id);
+            await repo.Create(order);
+
+            // Act
+            Result<NotificationOrderWithStatus, CancellationError> result = await repo.CancelOrder(order.Id, order.Creator.ShortName);
+
+            // Assert   
+            result.Match(
+                success =>
+                    throw new Exception("No success value should be returned if order is not found in database."),
+                error =>
+                {
+                    Assert.Equal(CancellationError.CancellationProhibited, error);
+                    return true;
+                });
+        }
+
+        [Fact]
+        public async Task CancelOrder_CancellationConditionSatisfied_ReturnsOrder()
+        {
+            // Arrange
+            OrderRepository repo = (OrderRepository)ServiceUtil
+                .GetServices(new List<Type>() { typeof(IOrderRepository) })
+                .First(i => i.GetType() == typeof(OrderRepository));
+
+            NotificationOrder order = new()
+            {
+                Id = Guid.NewGuid(),
+                Created = DateTime.UtcNow,
+                Creator = new("test"),
+                Templates = new List<INotificationTemplate>()
+                {
+                    new SmsTemplate("Altinn", "This is the body")
+                },
+                RequestedSendTime = DateTime.UtcNow.AddMinutes(20)
+            };
+
+            _orderIdsToDelete.Add(order.Id);
+            await repo.Create(order);
+
+            // Act
+            Result<NotificationOrderWithStatus, CancellationError> result = await repo.CancelOrder(order.Id, order.Creator.ShortName);
+
+            // Assert   
+            result.Match(
+                success =>
+                {
+                    Assert.Equal(OrderProcessingStatus.Cancelled, success.ProcessingStatus.Status);
+                    return true;
+                },
+                error => throw new Exception("No error value should be returned if order satisfies cancellation conditions."));
         }
     }
 }
